Move push and merge under annex popup
The main reason for having the push commands under the main magit
pushing menu was that they share options with the push commands because
they're just pure push commands. However, I think it's less surprising
to have everything under the main annex menu.

`git annex merge' was also under a non-annex menu, but it's not a pure
merge command and doesn't share options with `git merge`, so I'm not
sure why I put it under the merge menu.