Remove minor mode
Magit's popups are global, but magit-annex-mode tried to set them for
individual repos, resulting in a misbehaving minor mode. If magit-annex
was enabled or disabled in one buffer/repo, it would then be enabled or
disabled globally.

This commit removes the minor mode and defines autoloads for the popup
menu.

See magit/magit#1396.