Remove trailing information from unused buffer
The output of 'git annex unused' ends with

|   (To see where data was previously used, try: git log --stat -S'KEY')
|
|   To remove unwanted data: git-annex dropunused NUMBER
|
| ok

Delete this since it doesn't apply to magit-annex-unused-mode.

If there are partially transferred files, they show up after the unused
files.

|  Some partially transferred data exists in temporary files:
|      NUMBER  KEY
|      2       SHA256E-...
|
|    To remove unwanted data: git-annex dropunused NUMBER

With these changes, this information will also be deleted. Given that
it's probably rare to have partially transferred files and that these
lines couldn't be operated on because they weren't being processed
correctly into sections, I think this is fine.

Fixes #10.