Remove pushing commands
The annex pushing popup existed to avoid checking out the 'git-annex'
branch in order to push it. While the majority of git annex users
probably rely on the sync command and never push 'git-annex' manually,
users that prefer a centralized git repository [1] do push 'git-annex'.

With upstream improvements to pushing commands [2], it's possible to
push branches that aren't checked out, so the annex pushing popup is no
longer helpful.

[1] http://git-annex.branchable.com/tips/centralized_git_repository_tutorial/
[2] 0b709443b92c12868be4d709b28cf27ba8526f00