magit-annex-add: Don't skip hidden directories
By default, 'git annex add' skips dotfiles unless they're explicitly
listed.  For untracked directories that start with a dot, this means
that

  git annex add .untracked-subdir

will add the directory, while

  git annex add .untracked-subdir/

will silently skip the it.

Teach magit-annex-add about this quirk so that it doesn't silently
ignore the request to add ".untracked-subdir".  This isn't necessary
for the '(untracked list)' section because 'git ls-files --other'
lists individual files of the subdirectory.