magit-annex-command{,-topdir}: Adjust for upstream changes
As of Magit's b1b7798f (magit-shell-command-topdir: return only the
read command, 2017-04-03), calling magit-annex-command or
magit-annex-command-topdir fails because magit-git-command no longer
takes a directory argument.

Adjust the magit-git-command calls for this signature change.  Also,
use magit-read-shell-command to read the command so that the interface
is consistent with the changes introduced by Magit's #3034.

Fixes #16.