Make file commands more useful from Dired buffers
The two main things here are (1) taking defaults from the Dired buffer
and (2) updating the Dired listing after an action so that
git-annex.el fontification and type changes (in the case of lock and
unlock) are up to date.

At this point, executing a command from Dired always displays the
process buffer.  I think this is convenient because in Dired buffers,
unlike in Magit buffers, the process buffer isn't available with a
single key press.  But if that annoys some users, we can consider
adjusting this (e.g., showing the process buffer if there is an error
and otherwise letting asynchronous calls be hanled by the user's
magit-process-popup-time setting).

Re: #20