tests: Set up Travis
This is based off of Magit's .travis.yml.  The main change, aside from
installing Magit as a dependency, is that git-annex is downloaded.
Use the standalone git-annex tarball because the git-annex available
on trusty is very old (and would fail because it doesn't have an undo
command).

The matrix has been stripped down a bit from Magit's.  It might be
worth extending this (for example, to test the minimum required
version of Magit supported), but this will do for now.

Hopefully this will help spot whether there are any further issues
with flaky tests.

Re: #19