tests: Rework magit-annex-tests-wait
7d4a503 (tests: Clean up active process on tear down, 2018-07-08)
addressed one flavor of test flakiness, but the Travis tests show
another occasional failure [0], perhaps caused by 7d4a503.

I'm unable to trigger this error locally, but my guess is it has to do
with running magit-annex-tests-kill-process after magit-this-process
is set to nil but before magit-process-sentinel has finished.  Instead
of monitoring magit-this-process, let's wrap magit-process-sentinel
and set a flag when the sentinel has finished.

Re: #19

[0]: https://travis-ci.org/magit/magit-annex/jobs/402176949