unused: Do not override magit-log binding
magit-annex-unused-log invokes magit-log, while populating the -S flag
with the key value at point.  Doing so involves mucking around with
the log arguments in the transient, and, with recent changes in
Magit's handling of buffer/transient arguments, this approach would
need to be reworked.  But overriding magit-log to provide -S probably
wasn't a good idea to begin with because it couples
magit-annex-unused-log too closely with magit-log and automatically
populating or changing the value of -S might surprise users.

Instead adjust the command to be specific for the unused section, make
it show the log directly, and give it a dedicated binding.  This is
now a closer parallel to magit-dired-log and magit-log-buffer-file.