magit-annex-add: Note need for v6+ rework
In version 5 repositories, being "unlocked" is a transient,
between-commit state.  Unlocked files are a subset of all modified
files, so it was reasonable for magit-annex-add to add unlocked files
when called with point on the "Unstaged changes" section.  That's no
longer the case with v6+ repositories, where an unlocked file is a
file pointer that is tracked by git.