magit-annex-unlocked-files: Rewrite for v6+ compatibility
To get a list of unlocked files, we look at files that have a type
change difference between the working tree and the index.  For version
5 repositories, this should reliably give all unlocked files (but it
is susceptible to also returning false positives because non-annex
files can of course have a type change too).

For version 6 and above, however, this approach fails.  When a file is
unlocked, it changes to a pointer file for use with clean/smudge
filters, and this change will not show up in a diff of the working
tree and the index.  Also, the unlocked representation can persist
across commits.

As of git-annex version 7.20191009, we can use `git annex find
--unlocked` to get a list of the files.  Use this approach when we
can, falling back to the previous approach.  This unfortunately means
that magit-annex-unlocked-files won't return unlocked files in v6+
repos for git-annex versions before 7.20191009.

Fixes #23.