unused-log: Add magit-git-log compatibility kludge
One of the updates made in b083c72 (Adjust to the post refresh-args
world, 2019-04-20) was to adjust for the magit-git-log to
magit-log-setup-buffer rename.  No compatibility kludge was added
because in general magit-annex's master is tracking magit's.

However, a definition for magit-annex was recently submitted to Guix.
Guix has a magit built from c761d28, which is after the transient
switch but before the magit-git-log rename.  So, ideally the Guix
definition for magit-annex would use a magit-annex commit that
supports transient while not breaking magit-annex-unused-log due to
the rename.

Make that possible by adding a temporary kludge.