Custom regexp to match wildcard filenames
If a file contains "[" in its name, magit-annex is unable to process
it, because it is interpreted as a wildcard and the resulting filename
after expansion tends not to match any file. To handle such files I
found the regexp "[*?]" works well, at the expense of not being able
to use [] wildcards.

The original regexp of "[[.*+\\^$?]" seems superfluous when only sh
wildcards are supported, and the filename is shell quoted before being
passed to the shell anyway.